refactor: Rename providers for clarity, move timeline util out of providers/
monitor.py and site_monitoring.py were confusingly similar names for two
different things (bench-wide system/process metrics vs. one site's
request/job analytics). Rename explicitly:

- providers/monitor.py -> providers/system_monitoring.py
  MonitorProvider -> SystemMonitoringProvider
- providers/site_monitoring.py: SiteMonitoringProvider (unchanged, now
  reads unambiguously against its counterpart)

timeline.py (TimelinePoint/build_timeline) isn't shaped like a provider -
no bench_root/window, no log reading, just bucketing/ranking math - so it
moves to admin/backend/internal/, alongside the other stateless helpers
(jwks.py, rate_limiter.py) rather than living among the Provider classes.
